% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{k_steady_state}
\alias{k_steady_state}
\title{Steady State Capital Stock}
\usage{
k_steady_state(alpha, beta, delta)
}
\arguments{
\item{alpha}{Numeric. Capital share \eqn{\alpha} in production.}

\item{beta}{Numeric. Discount factor \eqn{\beta}.}

\item{delta}{Numeric. Depreciation rate of capital \eqn{\delta}.}
}
\value{
A \code{Numeric} with the following components:
}
\description{
This function computes the steady state capital stock of the standard
deterministic neoclassical growth model
}
\examples{
k_steady_state(alpha = 0.33, beta = 0.96, delta = 0.05)

}
