% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogro-package.R
\docType{package}
\name{neogro-package}
\alias{neogro-package}
\title{neogro: Neoclassical Growth Model}
\description{
Solve the deterministic neoclassical growth model.
}
\author{
\strong{Maintainer}: Stefan Moog \email{stefan.moog@prognos.com}

}
\keyword{internal}
