% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test-golden}
\alias{test-golden}
\alias{test_golden}
\alias{test_golden2}
\title{Test Golden Section Maximization}
\usage{
test_golden(ax = -2.2, bx = 0, cx = 10, tol = 0.001)

test_golden2(ax = -2.2, bx = 0, cx = 10, tol = 0.001)
}
\arguments{
\item{ax, bx, cx}{interval and starting value for golden section
search}

\item{tol}{tolerance for stopping criterion}
}
\value{
A \code{Numeric}
}
\description{
Test Golden Section Maximization
}
\examples{
test_golden()
}
